Final Commit
Ready for review. Added in a shadow around the body element on the left and right side. Added a border around the section on the left and made the corners round.
Added the background to the top of the page. Added the comment backgrounds behind the reviews. Added a perspective on the tabs and rotated them. Added a gradient background in the description box of the page. Set the corners of that box to round.